feat: using guards on comments routes

+++ src/modules/comments/comments.controller.ts
@@ -1,18 +1,40 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Req, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { CommentsService } from './comments.service';
 import { UpdateCommentDto } from './dto/update-comment.dto';
+import { JwtAuthGuard } from '../account/guards/jwt.guard';
 
 @Controller('api/comments')
 export class CommentsController {
 	constructor(private readonly commentsService: CommentsService) { }
 
 	@Patch(':id')
-	public async update(@Param('id') id: string, @Body() data: UpdateCommentDto) {
+	@UseGuards(JwtAuthGuard)
+	public async update(
+		@Param('id') id: string,
+		@Body() data: UpdateCommentDto,
+		@Req() req: any
+	) {
+		const findComment = await this.commentsService.findOneById(+id);
+
+		if (!(findComment.account_id === req.user.id)) {
+			throw new UnauthorizedException("You cant update other user's comments");
+		}
+
 		return this.commentsService.update(+id, data);
 	}
 
 	@Delete(':id')
-	public async remove(@Param('id') id: string) {
+	@UseGuards(JwtAuthGuard)
+	public async remove(
+		@Param('id') id: string,
+		@Req() req: any
+	) {
+		const findComment = await this.commentsService.findOneById(+id);
+
+		if (!(findComment.account_id === req.user.id)) {
+			throw new UnauthorizedException("You cant delete other user's comments");
+		}
+
 		return this.commentsService.remove(+id);
 	}
 }
